Moved post processing effects to its own file
