+++ src/CMakeLists.txt
@@ -1062,6 +1062,7 @@ set( FASTMATH_SOURCES
 	gl/renderer/gl_renderstate.cpp
 	gl/renderer/gl_renderbuffers.cpp
 	gl/renderer/gl_lightdata.cpp
+	gl/renderer/gl_postprocess.cpp
 	gl/hqnx/init.cpp
 	gl/hqnx/hq2x.cpp
 	gl/hqnx/hq3x.cpp
