+++ src/gl/scene/gl_scene.cpp
@@ -72,10 +72,6 @@
 #include "gl/scene/gl_drawinfo.h"
 #include "gl/scene/gl_portal.h"
 #include "gl/shaders/gl_shader.h"
-#include "gl/shaders/gl_bloomshader.h"
-#include "gl/shaders/gl_blurshader.h"
-#include "gl/shaders/gl_tonemapshader.h"
-#include "gl/shaders/gl_presentshader.h"
 #include "gl/stereo3d/gl_stereo3d.h"
 #include "gl/stereo3d/scoped_view_shifter.h"
 #include "gl/textures/gl_translate.h"
@@ -96,26 +92,8 @@ CVAR(Float, gl_mask_threshold, 0.5f,CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 CVAR(Float, gl_mask_sprite_threshold, 0.5f,CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 CVAR(Bool, gl_sort_textures, false, CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
 
-CVAR(Bool, gl_bloom, false, CVAR_ARCHIVE|CVAR_GLOBALCONFIG);
-CVAR(Float, gl_bloom_amount, 1.4f, 0)
-CVAR(Float, gl_exposure, 0.0f, 0)
-
-CUSTOM_CVAR(Int, gl_tonemap, 0, CVAR_ARCHIVE|CVAR_GLOBALCONFIG)
-{
-	if (self < 0 || self > 4)
-		self = 0;
-}
-
-CUSTOM_CVAR(Int, gl_bloom_kernel_size, 7, 0)
-{
-	if (self < 3 || self > 15 || self % 2 == 0)
-		self = 7;
-}
-
 EXTERN_CVAR (Bool, cl_capfps)
 EXTERN_CVAR (Bool, r_deathcamera)
-EXTERN_CVAR(Float, vid_brightness)
-EXTERN_CVAR(Float, vid_contrast)
 
 
 extern int viewpitch;
@@ -221,257 +199,6 @@ void FGLRenderer::Set3DViewport(bool toscreen)
 	glEnable(GL_STENCIL_TEST);
 	glStencilFunc(GL_ALWAYS,0,~0);	// default stencil
 	glStencilOp(GL_KEEP,GL_KEEP,GL_REPLACE);
-}
-
-//-----------------------------------------------------------------------------
-//
-// Adds bloom contribution to scene texture
-//
-//-----------------------------------------------------------------------------
-
-void FGLRenderer::BloomScene()
-{
-	// Only bloom things if enabled and no special fixed light mode is active
-	if (!gl_bloom || !FGLRenderBuffers::IsEnabled() || gl_fixedcolormap != CM_DEFAULT)
-		return;
-
-	const float blurAmount = gl_bloom_amount;
-	int sampleCount = gl_bloom_kernel_size;
-
-	// TBD: Maybe need a better way to share state with other parts of the pipeline
-	GLboolean blendEnabled, scissorEnabled;
-	GLint currentProgram, blendEquationRgb, blendEquationAlpha, blendSrcRgb, blendSrcAlpha, blendDestRgb, blendDestAlpha;
-	glGetBooleanv(GL_BLEND, &blendEnabled);
-	glGetBooleanv(GL_SCISSOR_TEST, &scissorEnabled);
-	glGetIntegerv(GL_CURRENT_PROGRAM, &currentProgram);
-	glGetIntegerv(GL_BLEND_EQUATION_RGB, &blendEquationRgb);
-	glGetIntegerv(GL_BLEND_EQUATION_ALPHA, &blendEquationAlpha);
-	glGetIntegerv(GL_BLEND_SRC_RGB, &blendSrcRgb);
-	glGetIntegerv(GL_BLEND_SRC_ALPHA, &blendSrcAlpha);
-	glGetIntegerv(GL_BLEND_DST_RGB, &blendDestRgb);
-	glGetIntegerv(GL_BLEND_DST_ALPHA, &blendDestAlpha);
-
-	glDisable(GL_BLEND);
-	glDisable(GL_SCISSOR_TEST);
-
-	const auto &level0 = mBuffers->BloomLevels[0];
-
-	// Extract blooming pixels from scene texture:
-	glBindFramebuffer(GL_FRAMEBUFFER, level0.VFramebuffer);
-	glViewport(0, 0, level0.Width, level0.Height);
-	mBuffers->BindSceneTexture(0);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-	mBloomExtractShader->Bind();
-	mBloomExtractShader->SceneTexture.Set(0);
-	mBloomExtractShader->Exposure.Set(mCameraExposure);
-	{
-		FFlatVertex *ptr = mVBO->GetBuffer();
-		ptr->Set(-1.0f, -1.0f, 0, 0.0f, 0.0f); ptr++;
-		ptr->Set(-1.0f, 1.0f, 0, 0.0f, 1.0f); ptr++;
-		ptr->Set(1.0f, -1.0f, 0, 1.0f, 0.0f); ptr++;
-		ptr->Set(1.0f, 1.0f, 0, 1.0f, 1.0f); ptr++;
-		mVBO->RenderCurrent(ptr, GL_TRIANGLE_STRIP);
-	}
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-
-	// Blur and downscale:
-	for (int i = 0; i < FGLRenderBuffers::NumBloomLevels - 1; i++)
-	{
-		const auto &level = mBuffers->BloomLevels[i];
-		const auto &next = mBuffers->BloomLevels[i + 1];
-		mBlurShader->BlurHorizontal(mVBO, blurAmount, sampleCount, level.VTexture, level.HFramebuffer, level.Width, level.Height);
-		mBlurShader->BlurVertical(mVBO, blurAmount, sampleCount, level.HTexture, next.VFramebuffer, next.Width, next.Height);
-	}
-
-	// Blur and upscale:
-	for (int i = FGLRenderBuffers::NumBloomLevels - 1; i > 0; i--)
-	{
-		const auto &level = mBuffers->BloomLevels[i];
-		const auto &next = mBuffers->BloomLevels[i - 1];
-
-		mBlurShader->BlurHorizontal(mVBO, blurAmount, sampleCount, level.VTexture, level.HFramebuffer, level.Width, level.Height);
-		mBlurShader->BlurVertical(mVBO, blurAmount, sampleCount, level.HTexture, level.VFramebuffer, level.Width, level.Height);
-
-		// Linear upscale:
-		glBindFramebuffer(GL_FRAMEBUFFER, next.VFramebuffer);
-		glViewport(0, 0, next.Width, next.Height);
-		glActiveTexture(GL_TEXTURE0);
-		glBindTexture(GL_TEXTURE_2D, level.VTexture);
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-		mBloomCombineShader->Bind();
-		mBloomCombineShader->BloomTexture.Set(0);
-		{
-			FFlatVertex *ptr = mVBO->GetBuffer();
-			ptr->Set(-1.0f, -1.0f, 0, 0.0f, 0.0f); ptr++;
-			ptr->Set(-1.0f, 1.0f, 0, 0.0f, 1.0f); ptr++;
-			ptr->Set(1.0f, -1.0f, 0, 1.0f, 0.0f); ptr++;
-			ptr->Set(1.0f, 1.0f, 0, 1.0f, 1.0f); ptr++;
-			mVBO->RenderCurrent(ptr, GL_TRIANGLE_STRIP);
-		}
-	}
-
-	mBlurShader->BlurHorizontal(mVBO, blurAmount, sampleCount, level0.VTexture, level0.HFramebuffer, level0.Width, level0.Height);
-	mBlurShader->BlurVertical(mVBO, blurAmount, sampleCount, level0.HTexture, level0.VFramebuffer, level0.Width, level0.Height);
-
-	// Add bloom back to scene texture:
-	mBuffers->BindSceneFB();
-	glViewport(0, 0, mOutputViewport.width, mOutputViewport.height);
-	glEnable(GL_BLEND);
-	glBlendEquation(GL_FUNC_ADD);
-	glBlendFunc(GL_ONE, GL_ONE);
-	glActiveTexture(GL_TEXTURE0);
-	glBindTexture(GL_TEXTURE_2D, level0.VTexture);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-	mBloomCombineShader->Bind();
-	mBloomCombineShader->BloomTexture.Set(0);
-	{
-		FFlatVertex *ptr = mVBO->GetBuffer();
-		ptr->Set(-1.0f, -1.0f, 0, 0.0f, 0.0f); ptr++;
-		ptr->Set(-1.0f, 1.0f, 0, 0.0f, 1.0f); ptr++;
-		ptr->Set(1.0f, -1.0f, 0, 1.0f, 0.0f); ptr++;
-		ptr->Set(1.0f, 1.0f, 0, 1.0f, 1.0f); ptr++;
-		mVBO->RenderCurrent(ptr, GL_TRIANGLE_STRIP);
-	}
-
-	if (blendEnabled)
-		glEnable(GL_BLEND);
-	if (scissorEnabled)
-		glEnable(GL_SCISSOR_TEST);
-	glBlendEquationSeparate(blendEquationRgb, blendEquationAlpha);
-	glBlendFuncSeparate(blendSrcRgb, blendDestRgb, blendSrcAlpha, blendDestAlpha);
-	glUseProgram(currentProgram);
-}
-
-//-----------------------------------------------------------------------------
-//
-// Tonemap scene texture and place the result in the HUD/2D texture
-//
-//-----------------------------------------------------------------------------
-
-void FGLRenderer::TonemapScene()
-{
-	if (gl_tonemap == 0)
-		return;
-
-	GLboolean blendEnabled, scissorEnabled;
-	glGetBooleanv(GL_BLEND, &blendEnabled);
-	glGetBooleanv(GL_SCISSOR_TEST, &scissorEnabled);
-
-	glDisable(GL_BLEND);
-	glDisable(GL_SCISSOR_TEST);
-
-	mBuffers->BindHudFB();
-	mBuffers->BindSceneTexture(0);
-	mTonemapShader->Bind();
-	mTonemapShader->SceneTexture.Set(0);
-	mTonemapShader->Exposure.Set(mCameraExposure);
-	
-	FFlatVertex *ptr = mVBO->GetBuffer();
-	ptr->Set(-1.0f, -1.0f, 0, 0.0f, 0.0f); ptr++;
-	ptr->Set(-1.0f, 1.0f, 0, 0.0f, 1.0f); ptr++;
-	ptr->Set(1.0f, -1.0f, 0, 1.0f, 0.0f); ptr++;
-	ptr->Set(1.0f, 1.0f, 0, 1.0f, 1.0f); ptr++;
-	mVBO->RenderCurrent(ptr, GL_TRIANGLE_STRIP);
-
-	if (blendEnabled)
-		glEnable(GL_BLEND);
-	if (scissorEnabled)
-		glEnable(GL_SCISSOR_TEST);
-}
-
-//-----------------------------------------------------------------------------
-//
-// Gamma correct while copying to frame buffer
-//
-//-----------------------------------------------------------------------------
-
-void FGLRenderer::Flush()
-{
-	if (FGLRenderBuffers::IsEnabled())
-	{
-		glDisable(GL_MULTISAMPLE);
-		glDisable(GL_DEPTH_TEST);
-		glDisable(GL_STENCIL_TEST);
-
-		mBuffers->BindOutputFB();
-
-		// Calculate letterbox
-		int clientWidth = framebuffer->GetClientWidth();
-		int clientHeight = framebuffer->GetClientHeight();
-		float scaleX = clientWidth / (float)mOutputViewport.width;
-		float scaleY = clientHeight / (float)mOutputViewport.height;
-		float scale = MIN(scaleX, scaleY);
-		int width = (int)round(mOutputViewport.width * scale);
-		int height = (int)round(mOutputViewport.height * scale);
-		int x = (clientWidth - width) / 2;
-		int y = (clientHeight - height) / 2;
-
-		// Black bars around the box:
-		glViewport(0, 0, clientWidth, clientHeight);
-		glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
-		glEnable(GL_SCISSOR_TEST);
-		if (y > 0)
-		{
-			glScissor(0, 0, clientWidth, y);
-			glClear(GL_COLOR_BUFFER_BIT);
-		}
-		if (clientHeight - y - height > 0)
-		{
-			glScissor(0, y + height, clientWidth, clientHeight - y - height);
-			glClear(GL_COLOR_BUFFER_BIT);
-		}
-		if (x > 0)
-		{
-			glScissor(0, y, x, height);
-			glClear(GL_COLOR_BUFFER_BIT);
-		}
-		if (clientWidth - x - width > 0)
-		{
-			glScissor(x + width, y, clientWidth - x - width, height);
-			glClear(GL_COLOR_BUFFER_BIT);
-		}
-		glDisable(GL_SCISSOR_TEST);
-
-		// Present what was rendered:
-		glViewport(x, y, width, height);
-
-		GLboolean blendEnabled;
-		GLint currentProgram;
-		glGetBooleanv(GL_BLEND, &blendEnabled);
-		glGetIntegerv(GL_CURRENT_PROGRAM, &currentProgram);
-		glDisable(GL_BLEND);
-
-		mPresentShader->Bind();
-		mPresentShader->InputTexture.Set(0);
-		if (framebuffer->IsHWGammaActive())
-		{
-			mPresentShader->Gamma.Set(1.0f);
-			mPresentShader->Contrast.Set(1.0f);
-			mPresentShader->Brightness.Set(0.0f);
-		}
-		else
-		{
-			mPresentShader->Gamma.Set(clamp<float>(Gamma, 0.1f, 4.f));
-			mPresentShader->Contrast.Set(clamp<float>(vid_contrast, 0.1f, 3.f));
-			mPresentShader->Brightness.Set(clamp<float>(vid_brightness, -0.8f, 0.8f));
-		}
-		mBuffers->BindHudTexture(0);
-
-		FFlatVertex *ptr = GLRenderer->mVBO->GetBuffer();
-		ptr->Set(-1.0f, -1.0f, 0, 0.0f, 0.0f); ptr++;
-		ptr->Set(-1.0f, 1.0f, 0, 0.0f, 1.0f); ptr++;
-		ptr->Set(1.0f, -1.0f, 0, 1.0f, 0.0f); ptr++;
-		ptr->Set(1.0f, 1.0f, 0, 1.0f, 1.0f); ptr++;
-		GLRenderer->mVBO->RenderCurrent(ptr, GL_TRIANGLE_STRIP);
-
-		if (blendEnabled)
-			glEnable(GL_BLEND);
-		glUseProgram(currentProgram);
-	}
 }
 
 //-----------------------------------------------------------------------------
